requirements.txt
PyYAML>=6.0
jsonschema>=4.0
requests>=2.0
beautifulsoup4>=4.0